refactor(runtime): extract issue run-task helper module (#1282)
